ML_SERVICE/requirements.txt
python-dotenv==1.0.1
pinecone==3.1.0
google-generativeai==0.4.0
langchain==0.1.10
langchain-google-genai==0.0.9
neo4j==5.17.0
fastapi==0.109.2
uvicorn==0.27.1
